requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
idna==3.4
importlib-metadata==6.4.1
multidict==6.0.4
numpy==1.21.6
outcome==1.2.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
